llm-experiments/iso-vector-db/requirements.txt
# Core dependencies
ollama>=0.1.0
faiss-cpu>=1.7.0
numpy>=1.21.0
python-dotenv>=1.0.0

# Optional: for index optimization
scikit-optimize>=0.9.0
